Fix controller tests. Disable broken ci flow.
