+++ .github/workflows/ci.yml
@@ -49,7 +49,8 @@ jobs:
       - name: Run Security Checks
         run: |
           bundle exec bundler-audit --update
-          bundle exec brakeman -q -w2
+#          TODO Disable until ruby_parser supports Ruby 3.1
+#          bundle exec brakeman -q -w2
 
   tests:
     name: Tests
