+++ test/factories/users.rb
@@ -31,6 +31,17 @@ FactoryBot.define do
     password { 'password' }
     # Default users to be confirmed so that tests may be performed using them
     confirmed_at { DateTime.now }
+    factory :user_with_full_recipes do
+      transient do
+        recipes_count { 5 }
+      end
+
+      after(:create) do |user, evaluator|
+        FactoryBot.create_list(:recipe_with_full_steps, evaluator.recipes_count, user: user)
+
+        user.reload
+      end
+    end
     factory :user_with_recipes do
       transient do
         recipes_count { 5 }
